requirements.txt
dateutils==0.6.6
MySQL-python==1.2.5
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
six==1.10.0